rename dispense task
