+++ hardhat.config.js
@@ -1,7 +1,7 @@
 const { join } = require('path');
 require('@nomicfoundation/hardhat-toolbox');
 require('./tasks/deploy.js');
-require('./tasks/dispence.js');
+require('./tasks/dispense.js');
 require('./tasks/clean.js');
 // require('hardhat-contract-sizer');
 const { mnemonic } = require('./.secret.json');
